Made the OpenGL includes work for mac which usues OpenGL/gl.h this could be changed to not directly include gl.h and use the Qt versions which will manage the cross platform includes for you.
changed a few uint defines to be unsigned int

Added a few virtual destructors to try and get rid of a build warning
especially arounf the GeometryOpDef files however this has not fixed it
on the mac build, this could be something to do with the BOOST export
macro.

Added a vcpkg.json file to allow local build of required packages.
However this is not working yet due to some incompatabilities. Will try
and update once I can find a decent baseline.

Updated OpDefs CMakeLists.txt to add boost libraries as was getting
linker errors on mac (IIRC mac builds a lib vs header only on other
platforms).